GNS/requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-import-export==4.0.7
lxml==5.2.2
opcua==0.98.13
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
